+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `/@username` routes for profile pages.
 - Allowed to attach image from clipboard.
 - Started using media limits provided by the backend.
+- Send avatar and banner media types to server when updating profile info.
 
 ### Changed